requirements.txt
pandas
scipy
sqlalchemy
flask
python-dotenv
pymssql
uvicorn
langchain
streamlit
pygwalker
ollama
groq
pymysql